Updated surefire and failsafe plugins to versions that detect jupiter tests. Added jupiter engine as dependency so that it can run the tests. Found one failing test, will fix that first for #6
