+++ src/test/resources/simplelogger.properties
@@ -1,0 +1,7 @@
+# Turning off Jetty logging.
+org.eclipse.jetty.util.log.class=org.eclipse.jetty.util.log.Slf4jLog
+org.eclipse.jetty.LEVEL=OFF
+
+# Default levels for the project
+org.slf4j.simpleLogger.defaultLogLevel=info
+org.slf4j.simpleLogger.log.net.twasi.obsremotejava=off
